Update hashJS.js
In the compileTemplate method, we construct the function without changing the original logic of the template.

In the update method, we use the call method to invoke the compiled function, setting its this context to this.currentData. This allows the function to access properties of the data object as if they're properties of this.